Rename pycairo.h to py3cairo.h also in the source
This used to be installed as either pycairo.h or py3cairo.h
depending on the Python version. Now that Python 2 is gone we can
simplify it and just use py3cairo.h everywhere.
